modulize server and router
